+++ src/app/admin/admin-routing.module.ts
@@ -1,5 +1,4 @@
 import {RouterModule, Routes} from "@angular/router";
-import {resolveProductWithRedirect} from "../shared/guards/product.resolver";
 import {NgModule} from "@angular/core";
 import {AdminDashboardPageComponent} from "./components/admin-dashboard-page/admin-dashboard-page.component";
 import {AdminProductsPageComponent} from "./components/admin-products-page/admin-products-page.component";
@@ -19,7 +18,6 @@ const routes: Routes = [
       {
         path: 'product/edit/:productID',
         component: AdminProductEditPageComponent,
-        ...resolveProductWithRedirect('/admin/products'),
       },
       { path: 'product/add', component: AdminProductCreatePageComponent },
       { path: 'products', component: AdminProductsPageComponent },