+++ src/app/product/services/product.repository.ts
@@ -5,6 +5,7 @@ import {HttpClient} from "@angular/common/http";
 import {Observable} from "rxjs";
 import {Environment} from "../../core/models/environment";
 import {ENV_TOKEN} from "../../app.module";
+import {delay} from "rxjs/operators";
 
 @Injectable({providedIn: 'root'})
 export class ProductRepository extends BaseRepository {
@@ -16,10 +17,12 @@ export class ProductRepository extends BaseRepository {
   }
 
   getProducts(): Observable<ProductModel[]> {
-    return this.sendGet<ProductModel[]>('/products');
+    return this.sendGet<ProductModel[]>('/products')
+      .pipe(delay(500));
   }
 
   getProduct(id: number): Observable<ProductModel> {
-    return this.sendGet<ProductModel>(`/products/${id}`);
+    return this.sendGet<ProductModel>(`/products/${id}`)
+      .pipe(delay(500));
   }
 }