+++ src/app/shared/guards/is-empty-cart.guard.ts
@@ -1,16 +1,25 @@
 import {ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree} from "@angular/router";
-import {CartService} from "../../cart/services/cart.service";
 import {Injectable} from "@angular/core";
+import {Store} from "@ngrx/store";
+import {AppState} from "../../store/app.reducer";
+import {map, take} from "rxjs/operators";
+import {Observable} from "rxjs";
 
 @Injectable({providedIn: 'root'})
 export class IsEmptyCartGuard implements CanActivate {
-  constructor(private cartService: CartService, private router: Router) {}
+  constructor(private store: Store<AppState>, private router: Router) {}
 
-  canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean | UrlTree {
-    if (!this.cartService.isEmptyCart()) {
-      return true;
-    }
+  canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean | UrlTree> {
+    return this.store.select('cart')
+      .pipe(
+        take(1),
+        map(data => {
+          if (!data.isEmpty) {
+            return true;
+          }
 
-    return this.router.parseUrl('/products');
+          return this.router.parseUrl('/products');
+        }),
+      );
   }
 }